+++ src/index-dev.ts
@@ -3,7 +3,7 @@ import { Sustainability } from ".";
 (async () => {
   Sustainability.auditStream.pipe(process.stdout);
   const report = await Sustainability.audit("https://reddit.com", {
-    connectionSettings: { streams: false, coldRun: false },
+    connectionSettings: { streams: true, coldRun: false },
   });
 
   console.log(JSON.stringify(report));