+++ src/types/audit.ts
@@ -137,3 +137,12 @@ export interface SkipMeta {
 }
 
 export type AuditType = Result | SkipResult | undefined;
+
+export interface AuditStreamChunk {
+  meta: {
+    id?:string,
+    status: "bootstrap" | "collect" | "audit"
+    total: number
+  }
+  audit?: AuditType
+}